bigInteger
This file implements members of a class for arbitrary-size integer arithmetic. It is meant to help get around the maximum value limit for int.

The program parses each equation for correctness. If the string is not a valid equation then it will not be processed.

The program has proven to handle complex equations (documents over over 100 thousand characters long) in fractions of a second.